fix magnitude type

+++ tygle_docs/types/resources/models/dimension.py
@@ -3,5 +3,5 @@ from tygle_docs.types.enums import Unit
 
 
 class Dimension(BaseModel):
-    magnitude: int = Field()
+    magnitude: float = Field()
     unit: Unit = Field()
